errorwindow hotkeys

+++ BugTracking/BugTracking/Users.cs
@@ -112,36 +112,15 @@ namespace BugTracking
         }
 
         private void Users_KeyDown(object sender, KeyEventArgs e)
+
         {
-            if (e.Modifiers == Keys.Control && e.KeyCode == Keys.A)
-            {
-                AddUser adduser = new AddUser();
-                adduser.Show();
-
-            }
-            if (e.Modifiers == Keys.Control && e.KeyCode == Keys.E)
-            {
-                if (dataGridView1.SelectedRows.Count > 0)
+                if (e.Modifiers == Keys.Control && e.KeyCode == Keys.A)
                 {
-                    int index = dataGridView1.SelectedRows[0].Index;
-                    int id = 0;
-                    bool converted = Int32.TryParse(dataGridView1[0, index].Value.ToString(), out id);
-                    if (converted == false)
-                        return;
-                    EditUser edituser = new EditUser(id);
-                    edituser.Show();
+                    AddUser adduser = new AddUser();
+                    adduser.Show();
+
                 }
-
-            }
-            if (e.KeyCode == Keys.F5)
-            {
-                BTContext db = new BTContext();
-                dataGridView1.DataSource = db.Users.ToList();
-
-            }
-            if (e.KeyCode == Keys.Delete)
-            {
-                using (BTContext db = new BTContext())
+                if (e.Modifiers == Keys.Control && e.KeyCode == Keys.E)
                 {
                     if (dataGridView1.SelectedRows.Count > 0)
                     {
@@ -150,15 +129,43 @@ namespace BugTracking
                         bool converted = Int32.TryParse(dataGridView1[0, index].Value.ToString(), out id);
                         if (converted == false)
                             return;
-                        User user = db.Users.Find(id);
-                        db.Users.Remove(user);
-                        db.SaveChanges();
+                        EditUser edituser = new EditUser(id);
+                        edituser.Show();
+                    }
+
+                }
+                if (e.KeyCode == Keys.F5)
+                {
+                    BTContext db = new BTContext();
+                    dataGridView1.DataSource = db.Users.ToList();
+
+                }
+                if (e.KeyCode == Keys.Delete)
+                {
+                    using (BTContext db = new BTContext())
+                    {
+                        if (dataGridView1.SelectedRows.Count > 0)
+                        {
+                            int index = dataGridView1.SelectedRows[0].Index;
+                            int id = 0;
+                            bool converted = Int32.TryParse(dataGridView1[0, index].Value.ToString(), out id);
+                            if (converted == false)
+                                return;
+                            User user = db.Users.Find(id);
+                            db.Users.Remove(user);
+                            db.SaveChanges();
+                        }
                     }
                 }
-            }
 
+            
         }
+         
+           
     }
        
 }
+    
+       
+
